# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/role/v1/list": {
            "get": {
                "description": "get all roles",
                "tags": [
                    "roles"
                ],
                "summary": "receiving roles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseRoles"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    }
                }
            }
        },
        "/api/ticket-response/v1/list": {
            "get": {
                "description": "get all ticketResponses",
                "tags": [
                    "ticketResponses"
                ],
                "summary": "receiving ticketResponses",
                "parameters": [
                    {
                        "type": "string",
                        "description": "9999990000",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-01-25T10:30:00.000Z",
                        "name": "createdAtFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-02-25T10:30:00.000Z",
                        "name": "createdAtTo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-01-25T10:30:00.000Z",
                        "name": "updatedAtFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-02-25T10:30:00.000Z",
                        "name": "updatedAtTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseTicketResponses"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    }
                }
            }
        },
        "/api/ticket-response/v1/my-list": {
            "get": {
                "description": "get all ticketResponses for user id",
                "tags": [
                    "ticketResponses"
                ],
                "summary": "receiving ticketResponses",
                "parameters": [
                    {
                        "type": "string",
                        "description": "9999990000",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-01-25T10:30:00.000Z",
                        "name": "createdAtFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-02-25T10:30:00.000Z",
                        "name": "createdAtTo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-01-25T10:30:00.000Z",
                        "name": "updatedAtFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2019-02-25T10:30:00.000Z",
                        "name": "updatedAtTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseTicketResponses"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    }
                }
            }
        },
        "/api/ticket-response/v1/{id}": {
            "get": {
                "description": "Get a specific ticketResponse by ID",
                "tags": [
                    "ticketResponses"
                ],
                "summary": "Get TicketResponse by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TicketResponse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseTicketResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    }
                }
            }
        },
        "/api/ticketResponse/v1/create": {
            "post": {
                "description": "Create a specific ticketResponse",
                "tags": [
                    "ticketResponses"
                ],
                "summary": "Create TicketResponses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.BadResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.BadResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "domain.Response": {
            "type": "object",
            "properties": {
                "payload": {},
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "domain.ResponseRoles": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Role"
                    }
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "domain.ResponseTicketResponse": {
            "type": "object",
            "properties": {
                "payload": {
                    "$ref": "#/definitions/domain.TicketResponse"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "domain.ResponseTicketResponses": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TicketResponse"
                    }
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "domain.Role": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "name": {
                    "$ref": "#/definitions/domain.RoleName"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "domain.RoleName": {
            "type": "string",
            "enum": [
                "user",
                "manager"
            ],
            "x-enum-varnames": [
                "RoleNameUser",
                "RoleNameManager"
            ]
        },
        "domain.TicketResponse": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "applicationId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "managerId": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    }
}